back_v2/requirements.txt
aioredis==1.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2025.1.31
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.1
constantly==23.10.4
cryptography==44.0.2
daphne==3.0.2
Django==5.2
django-channels==0.7.0
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
hiredis==3.1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
inflection==0.5.1
msgpack==1.1.0
oauthlib==3.2.2
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
setuptools==78.1.0
six==1.17.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
zope.interface==7.2
